refactor(commit_round): fold candidate selection into the validation service
The RFC's "commit validation service" covers both validating candidates and selecting the winner by priority. Move rank_applicable_candidates and compare_candidate_priority (rewritten with Ordering::then_with) from round into validate, so it owns the whole service; round.rs is left as the public surface and finalize orchestration.
